wan/video_api/requirements.txt
torch>=2.4.0
diffusers @ git+https://github.com/huggingface/diffusers.git
transformers>=4.40.0
accelerate>=0.30.0
sentencepiece
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
huggingface_hub
safetensors
imageio
imageio-ffmpeg
optimum-quanto
pillow
numpy
